=htj move directive tests into own package
