+++ akka-http-tests/src/test/java/akka/http/javadsl/server/directives/AuthenticationDirectivesTest.java
@@ -2,7 +2,7 @@
  * Copyright (C) 2009-2014 Typesafe Inc. <http://www.typesafe.com>
  */
 
-package akka.http.javadsl.server;
+package akka.http.javadsl.server.directives;
 
 import org.junit.Test;
 import scala.Option;
@@ -10,6 +10,7 @@ import scala.concurrent.Future;
 import akka.http.javadsl.model.HttpRequest;
 import akka.http.javadsl.model.headers.Authorization;
 import akka.http.javadsl.testkit.*;
+import akka.http.javadsl.server.*;
 import static akka.http.javadsl.server.Directives.*;
 
 public class AuthenticationDirectivesTest extends JUnitRouteTest {
